Cleaned up namespaces. Removed unused objects.
