--- AASAXUtilLib/AX/Framework/Tools/ModelManagement/ExecutionStatus.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Microsoft.Dynamics.AX.Framework.Tools.ModelManagement
-{
-    [Serializable]
-    public enum ExecutionStatus
-    {
-        Ok,
-        Error,
-        Warning,
-    }
-}
