--- AASAXUtilLib/AX/Framework/Tools/ModelManagement/ParameterConsoleCommandConstants.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.Dynamics.AX.Framework.Tools.ModelManagement
-{
-    public static class ParameterConsoleCommandConstants
-    {
-        public const string GenLicense = "GENLICENSE";
-    }
-}
