--- AASAXUtilLib/AX/Framework/Tools/ModelManagement/ParameterConsoleConstants.cs
@@ -1,8 +1,0 @@
-﻿namespace Microsoft.Dynamics.AX.Framework.Tools.ModelManagement
-{
-    public static class ParameterConsoleConstants
-    {
-        public const string NameValueSeparator = ":";
-        public const string ParameterSwitchDashIdentifier = "-";
-    }
-}
