+++ AASAXUtilLib/Properties/Resources.cs
